refactor: build scene root in entity pipeline
