+++ packages/loader/src/gltf/parser/GLTFEntityParser.ts
@@ -21,7 +21,13 @@ export class GLTFEntityParser extends GLTFParser {
 
     if (!promise) {
       if (index === undefined) {
-        promise = Promise.all(nodes.map((entityInfo, index) => this._parserSingleEntity(context, entityInfo, index)));
+        promise = Promise.all(
+          nodes.map((entityInfo, index) => this._parserSingleEntity(context, entityInfo, index))
+        ).then((entities) => {
+          this._buildEntityTree(context, entities);
+          this._createSceneRoots(context, entities);
+          return entities;
+        });
       } else {
         promise = this._parserSingleEntity(context, nodes[index], index);
       }
@@ -57,5 +63,50 @@ export class GLTFEntityParser extends GLTFParser {
     }
 
     return Promise.resolve(entity);
+  }
+
+  private _buildEntityTree(context: GLTFParserContext, entities: Entity[]): void {
+    const nodes = context.glTF.nodes;
+
+    for (let i = 0; i < nodes.length; i++) {
+      const { children } = nodes[i];
+      const entity = entities[i];
+
+      if (children) {
+        for (let j = 0; j < children.length; j++) {
+          const childEntity = entities[children[j]];
+
+          entity.addChild(childEntity);
+        }
+      }
+    }
+  }
+
+  private _createSceneRoots(context: GLTFParserContext, entities: Entity[]): void {
+    const { glTFResource, glTF } = context;
+    const { scene: sceneID = 0, scenes } = glTF;
+
+    if (!scenes) return;
+
+    const sceneRoots: Entity[] = [];
+
+    for (let i = 0; i < scenes.length; i++) {
+      const { nodes } = scenes[i];
+
+      if (!nodes) continue;
+
+      if (nodes.length === 1) {
+        sceneRoots[i] = entities[nodes[0]];
+      } else {
+        const rootEntity = new Entity(glTFResource.engine, "GLTF_ROOT");
+        for (let j = 0; j < nodes.length; j++) {
+          rootEntity.addChild(entities[nodes[j]]);
+        }
+        sceneRoots[i] = rootEntity;
+      }
+    }
+
+    glTFResource.sceneRoots = sceneRoots;
+    glTFResource.defaultSceneRoot = sceneRoots[sceneID];
   }
 }
